Revert "Refactored to use Decimal instead of double."
This reverts commit 874da3a57be8ff83ac8d65de59033f579ee7b0bc.

+++ lib/src/utils/utilities.dart
@@ -52,7 +52,7 @@ class _Utilities {
     this.settings = settings ?? MoneyFormatterSettings();
   }
 
-  Decimal amount;
+  double amount;
 
   MoneyFormatterSettings settings;
